scripts/eval_pipeline/requirements.txt
openai
google-genai
anthropic
mistralai
hydra-core
omegaconf
cloudscraper
beautifulsoup4
lxml
python-dotenv
pyyaml
